+++ test/controllers/dashboard_controller_test.rb
@@ -15,6 +15,10 @@ class DashboardControllerTest < ActionController::TestCase
   def dropdown_list(title)
     css_select("li.dropdown[title='#{title}'] ul")
   end
+  
+  def dropdown_link(order)
+    ".navbar-collapse > .nav li.dropdown:nth-of-type(#{order}) a"
+  end
 
   # given a dropdown list, return the list items as an array of strings
   # with symbols for header or divider
@@ -133,6 +137,71 @@ class DashboardControllerTest < ActionController::TestCase
     assert_select "nav a[href='#{batch_connect_sessions_path}']", 0
   end
 
+  #test whitelist mode 
+  test "should not create app menus if NavConfig.categories is empty and whitelist is enabled" do
+    SysRouter.stubs(:base_path).returns(Rails.root.join("test/fixtures/sys_with_gateway_apps"))
+    OodAppkit.stubs(:clusters).returns(OodCore::Clusters.load_file("test/fixtures/config/clusters.d"))
+    NavConfig.stubs(:categories_whitelist?).returns(true)
+    NavConfig.stubs(:categories).returns([])
+    
+    get :index
+    assert_response :success
+    assert_select ".navbar-collapse > .nav li.dropdown[title]", 0
+    NavConfig.unstub(:categories_whitelist?)
+    NavConfig.unstub(:categories)
+  end
+
+  test "should exclude gateway apps if NavConfig.categories is set to default and whitelist is enabled" do
+    SysRouter.stubs(:base_path).returns(Rails.root.join("test/fixtures/sys_with_gateway_apps"))
+    OodAppkit.stubs(:clusters).returns(OodCore::Clusters.load_file("test/fixtures/config/clusters.d"))
+    NavConfig.stubs(:categories_whitelist?).returns(true)
+    NavConfig.stubs(:categories).returns(["Files", "Jobs", "Clusters", "Interactive Apps"])
+
+    get :index
+    assert_response :success
+    assert_select ".navbar-collapse > .nav li.dropdown[title='Gateway Apps']", 0
+    NavConfig.unstub(:categories_whitelist?)
+    NavConfig.unstub(:categories)
+  end
+
+  test "should include gateway and System Installed apps in alphabetical order in end if NavConfig.categories is set to default listing and whitelist is disabled" do
+    SysRouter.stubs(:base_path).returns(Rails.root.join("test/fixtures/sys_with_gateway_apps"))
+    OodAppkit.stubs(:clusters).returns(OodCore::Clusters.load_file("test/fixtures/config/clusters.d"))
+    NavConfig.stubs(:categories_whitelist?).returns(false)
+    NavConfig.stubs(:categories).returns(["Files", "Jobs", "Clusters", "Interactive Apps"])
+    
+    get :index
+    assert_response :success
+    assert_select ".navbar-collapse > .nav li.dropdown[title]", 6
+    assert_select  dropdown_link(1), text: "Files"
+    assert_select  dropdown_link(2), text: "Jobs"
+    assert_select  dropdown_link(3), text: "Clusters"
+    assert_select  dropdown_link(4), text: "Interactive Apps"
+    assert_select  dropdown_link(5), text: "Gateway Apps"
+    assert_select  dropdown_link(6), text: "System Installed Apps"
+    NavConfig.unstub(:categories_whitelist?)
+    NavConfig.unstub(:categories)
+  end
+
+  test "should include all apps in alphabetical order if NavConfig.categories is set to nil or empty array and whitelist is disabled" do
+    SysRouter.stubs(:base_path).returns(Rails.root.join("test/fixtures/sys_with_gateway_apps"))
+    OodAppkit.stubs(:clusters).returns(OodCore::Clusters.load_file("test/fixtures/config/clusters.d"))
+    NavConfig.stubs(:categories_whitelist?).returns(false)
+    NavConfig.stubs(:categories).returns([])
+
+    get :index
+    assert_response :success
+    assert_select ".navbar-collapse > .nav li.dropdown[title]", 6
+    assert_select dropdown_link(1), text: "Clusters"
+    assert_select dropdown_link(2), text: "Files"
+    assert_select dropdown_link(3), text: "Gateway Apps"
+    assert_select dropdown_link(4), text: "Interactive Apps"
+    assert_select dropdown_link(5), text: "Jobs"
+    assert_select dropdown_link(6), text: "System Installed Apps"
+    NavConfig.unstub(:categories_whitelist?)
+    NavConfig.unstub(:categories)
+  end
+  
   test "should not create any empty links" do
     get :index