+++ CHANGELOG.md
@@ -6,6 +6,9 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Added whitelist mode to manage categories in the navbar.
+  [#295](https://github.com/OSC/ood-dashboard/issues/295)
 
 ## [1.27.2] - 2018-09-14
 ### Changed